requirements.txt
python_dotenv
azure-identity
azure-communication-identity
Flask